+++ src/Extension/DinkyTags.php
@@ -25,7 +25,9 @@ use TheLoom\Plugin\System\DinkyTags\Helper\ImageResolver;
 use TheLoom\Plugin\System\DinkyTags\Helper\MetaWriter;
 use TheLoom\Plugin\System\DinkyTags\Helper\Text;
 
+// phpcs:disable PSR1.Files.SideEffects
 \defined('_JEXEC') or die;
+// phpcs:enable PSR1.Files.SideEffects
 
 /**
  * Emits Open Graph and Twitter Card meta tags for frontend pages.